% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf_data.R
\docType{data}
\name{maf}
\alias{maf}
\title{SNP minor allele frequencies}
\format{a numeric vector}
\usage{
data(maf)
}
\description{
\code{maf} contains the minor allele frequencies for the 450 SNPs in
the \code{\link{genotypes}} dataset
}
\examples{
data(maf)
all((maf <= 0.5) & (maf >= 0))
}
\keyword{datasets}
