% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_data.R
\docType{data}
\name{propensity}
\alias{propensity}
\title{propensity scores}
\format{a numeric vector}
\usage{
data(propensity)
}
\description{
To obtain the scores for the samples in \code{\link{genotypes}},
we first remove all SNPs in the genomic interval
[18038910, 18288361] to alleviate linkage disequilibrium around the
target SNP rs2535708 (position 18184169). Afterwards, we apply
\code{\link{fast_HMM}} followed by \code{\link{cond_prob}}.
The results are stored in \code{propensity}.
}
\examples{
data(propensity)
}
\keyword{datasets}
