% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.R
\name{forward}
\alias{forward}
\title{Applies the forward algorithm to a genotype dataset}
\usage{
forward(X, p_init, p_trans, p_emit, ncores = 1)
}
\arguments{
\item{X}{genotype matrix. Each row corresponds to a separate sample}

\item{p_init}{marginal distributions for the first hidden state}

\item{p_trans}{3D dimensional array for the transition probabilities}

\item{p_emit}{3D dimensional array for the emission probabilities}

\item{ncores}{number of threads (default 1)}
}
\value{
A vector of log probabilities
}
\description{
Applies the \code{forward_sample} function to each row in \code{X}. If
the \code{ncores} > 1, the function calling is performed in a parallel
fashion to reduce the running time. The parallelization backend
is \code{doSNOW}.  If the latter package is not installed,
the function switches back to single-core mode.
}
\references{
Rabiner, Lawrence R. 'A tutorial on hidden Markov models
and selected applications in speech recognition.' Proceedings of the
IEEE 77.2 (1989): 257-286.
}
