% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sim_phenotype}
\alias{sim_phenotype}
\title{Simulates a binary phenotye}
\usage{
sim_phenotype(X, causal, model, intercept = TRUE)
}
\arguments{
\item{X}{genotype matrix}

\item{causal}{causal SNPs.}

\item{model}{disease model}

\item{intercept}{binary flag. If \code{intercept=TRUE}, a non-null intercept
is added so that the output is (approximately) balanced between cases and
controls.}
}
\value{
A vector of simulated phenotypes which are encoded as a two-level
  factor (TRUE/FALSE).
}
\description{
The phenotypes are simulated according to a logistic regression model.
Depending on the chosen configuration in \code{\link{sample_SNP}}, the model
includes different effect types: synergistic effects with the target,
marginal effects and additional epistatic effects. We offer the option to
generate a balanced phenotype vector between cases and controls, through the
\code{intercept} parameter.
}
\seealso{
\code{\link{sample_SNP}} and \code{\link{gen_model}}
}
